Change import and component convention
