+++ bmh_lims/frontend/src/assets/icons/index.js
@@ -1,0 +1,4 @@
+export { default as Logo } from './logo.svg'
+export { default as SampleIcon } from './samples.svg'
+export { default as StatusIcon } from './status.svg'
+export { default as WorkflowsIcon} from './workflows.svg'
