--- bmh_lims/frontend/src/assets/index.js
@@ -1,5 +1,0 @@
-export { default as Logo } from './icons/logo.svg'
-export { default as SampleIcon } from './icons/samples.svg'
-export { default as StatusIcon } from './icons/status.svg'
-export { default as WorkflowsIcon} from './icons/workflows.svg'
-export { csvReader as csvReader, xlsxReader } from './utils/csvReader'
