+++ bmh_lims/frontend/src/components/IconButton/index.js
@@ -1,20 +1,5 @@
 import React from 'react'
-import styled from 'styled-components'
-
-const StyledButton = styled.button`
-    width: ${props => props.width || 'auto'};
-    background-color: Transparent;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border: None;
-    ${props => props.animation}
-`
-
-const IconContainer = styled.div`
-    width: ${props => props.width || '50px'};
-    height: ${props => props.height || '50px'};
-`
+import { StyledButton, IconContainer } from './Styles'
 
 const IconButton = ({text, IconSVG, animation, iconProps, onClickHandler}) => {
     return (
