+++ bmh_lims/frontend/src/components/index.js
@@ -1,8 +1,7 @@
-export { default as Button } from './Button'
-export { default as IconButton } from './IconButton'
-export { default as IconLink } from './IconLink'
-export { default as Table } from './Table'
-export { default as CombinedLogo } from './Logo'
-
-export { StyledButton } from './Button'
-export { default as LinkButton } from './LinkButton'
+export { default as Button, StyledButton } from './Button/index'
+export { default as IconButton } from './IconButton/index'
+export { default as IconLink } from './IconLink/index'
+export { default as Table } from './Table/index'
+export { default as CombinedLogo } from './Logo/index'
+export { default as LinkButton } from './LinkButton/index'
+export { default as FileInputButton } from './FileInputButton/index'
