+++ bmh_lims/frontend/src/pages/Home/index.js
@@ -1,48 +1,7 @@
 import React from 'react'
-import styled, { keyframes, css } from 'styled-components'
 import { IconButton, IconLink, CombinedLogo } from 'components'
 import { SampleIcon, StatusIcon, WorkflowsIcon } from 'icons'
-
-// will worry about more responsiveness another time
-
-const ButtonBar = styled.div`
-    display: flex;
-    justify-content: space-between;
-    width: 50%;
-    margin-top: 25px; 
-`
-
-const Page = styled.div`
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    align-items: center;
-    justify-content: center;
-`
-
-// keeping this here for now since it's just one animation
-const appear = keyframes`
-    from{
-        transform: translate(0px, 5px);
-        opacity: 0;
-    }
-    to{
-        transform: translate(0px, 0px);
-        opacity: 1;
-    }
-`
-
-const buttonAnimation = css`
-    animation: ${appear} 0.5s linear 1;
-    &:hover {
-        transition: transform 0.5s;
-        transform: scale(1.2);
-    }
-    &:active {
-        transition: transform 0s;
-        transform: scale(1);
-    }
-`
+import { ButtonBar, Page, appear, buttonAnimation } from './Styles'
 
 // style props are kept as strings to retain both unit measures and quantity
 const HomePage = () => {
