+++ bmh_lims/frontend/src/pages/UploadSamples/Styles.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components'
+
+export const HeaderBar = styled.div`
+    display: flex;
+    justify-content: space-between;
+`
+
+export const PageContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    padding: 10px;
+`
+
+export const FooterBar = styled.div`
+    display: flex;
+    justify-content: right;
+`
