+++ bmh_lims/frontend/src/pages/UploadSamples/index.js
@@ -1,28 +1,28 @@
 import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
-import { Table, CombinedLogo, StyledButton, Button, LinkButton } from 'components'
-import styled, { css } from 'styled-components'
-import { csvReader, xlsxReader } from 'icons'
+import { Table, CombinedLogo, Button, LinkButton, FileInputButton } from 'components'
+import { HeaderBar, PageContainer, FooterBar } from './Styles'
+import { csvReader, xlsxReader } from 'utils'
 
-const csvToTable = (text, updateContent) => {
-    const lines = text.trim().split('\n').map(line => 
+const displayInTable = (dataText, updateContent) => {
+    const lines = dataText.trim().split('\n').map(line => 
         line.split(','))
     updateContent({headers: lines[0], content: lines.slice(1, lines.length)})
 }
 
 const onClickHandleUpload = (event, updateSubmittedFile, updateContent) => {
     event.preventDefault()
-    if(RegExp('\.[csv|tsv|xls|xlsx]$').test(event.target.files[0].name)) {
+    if(RegExp('\.[csv|xls|xlsx]$').test(event.target.files[0].name)) {
         const submittedFile = event.target.files[0]
         updateSubmittedFile({name: submittedFile.name, file: submittedFile})
 
         if(RegExp('\.[csv]$').test(submittedFile.name)) {
-            csvReader(submittedFile, (text) => csvToTable(text, updateContent))
+            csvReader(submittedFile, (dataText) => displayInTable(dataText, updateContent))
         } else if (RegExp('\.[xlsx|xlsx]$').test(submittedFile.name)) {
-            xlsxReader(submittedFile, (text) => csvToTable(text, updateContent))
+            xlsxReader(submittedFile, (dataText) => displayInTable(dataText, updateContent))
         }
     } else {
-        console.log('invalid file type')
+        console.log('invalid file type') // TODO: Toast
     }
 }
 
@@ -35,49 +35,13 @@ const onClickSubmit = (event, submittedFile) => {
     }
 }
 
-const FileInputContainer = styled.div`
-    position: relative;
-    width: 50px;
-`
-
-const FileInput = styled.input`
-    position: relative;
-    opacity: 0;
-    width: 100px;
-    z-index: 2;
-`
-const FileInputButton = styled(StyledButton)`
-    position: absolute;
-    left: 0;
-    z-index: 1;
-`
-
-const HeaderBar = styled.div`
-    display: flex;
-    justify-content: space-between;
-`
-
-const PageContainer = styled.div`
-    display: flex;
-    flex-direction: column;
-    padding: 10px;
-`
-
-const FooterBar = styled.div`
-    display: flex;
-    justify-content: right;
-`
-
 const UploadSamplesPage = () => {
     const [submittedFile, updateSubmittedFile] = useState({type: '', file: null})
     const [content, updateContent] = useState({headers: ['AAA', 'BBB', 'CCC', 'DDD', 'EEE'], content: [...Array(10).keys()].map((item, idx) => ['aaa', 'bbb', 'ccc', 'ddd', 'eee'])})
     return (
         <PageContainer>
             <HeaderBar>
-                <FileInputContainer>
-                    <FileInput type='file' onChange={(e) => onClickHandleUpload(e, updateSubmittedFile, updateContent)} />
-                    <FileInputButton>upload</FileInputButton>
-                </FileInputContainer>
+                <FileInputButton onChangeHandler={(e) => onClickHandleUpload(e, updateSubmittedFile, updateContent)} />
                 <Link to='/lims'><CombinedLogo height='50px' width='50px' /></Link>
             </HeaderBar>
             <Table headers={content.headers} content={content.content} />
