+++ bmh_lims/frontend/src/pages/index.js
@@ -1,2 +1,2 @@
-export { default as HomePage } from './Home'
-export { default as UploadSamplesPage } from './UploadSamples'
+export { default as HomePage } from './Home/index'
+export { default as UploadSamplesPage } from './UploadSamples/index'
